+++ lib/plausible_web/controllers/api/stats_controller.ex
@@ -567,12 +567,7 @@ defmodule PlausibleWeb.Api.StatsController do
 
   def countries(conn, params) do
     site = conn.assigns[:site]
-
-    query =
-      Query.from(site, params)
-      |> Filters.add_prefix()
-      |> Query.put_filter("visit:country", {:is_not, "\0\0"})
-
+    query = site |> Query.from(params) |> Filters.add_prefix()
     pagination = parse_pagination(params)
 
     countries =
@@ -624,12 +619,7 @@ defmodule PlausibleWeb.Api.StatsController do
 
   def regions(conn, params) do
     site = conn.assigns[:site]
-
-    query =
-      Query.from(site, params)
-      |> Filters.add_prefix()
-      |> Query.put_filter("visit:region", {:is_not, ""})
-
+    query = site |> Query.from(params) |> Filters.add_prefix()
     pagination = parse_pagination(params)
 
     regions =
@@ -662,12 +652,7 @@ defmodule PlausibleWeb.Api.StatsController do
 
   def cities(conn, params) do
     site = conn.assigns[:site]
-
-    query =
-      Query.from(site, params)
-      |> Filters.add_prefix()
-      |> Query.put_filter("visit:city", {:is_not, 0})
-
+    query = site |> Query.from(params) |> Filters.add_prefix()
     pagination = parse_pagination(params)
 
     cities =